Major changes, please read description.
In order to use this latest update, your database is required to be
cleared/deleted.

I have added a new column to the customer and business owner table
which will help us identify a username as either a customer or business
owner.

The ABS has been changed to help capture these changes, but the DB you
currently hold needs to be wiped.

Occasionally I will get “Connection failed to close” when I run ABS.